+++ projects/ngneat/transloco-utils/src/lib/types.ts
@@ -4,6 +4,9 @@ export interface TranslocoConfig {
   scopePathMap?: { [key: string]: string };
   langs?: string[];
   keysManager?: {
+    input?: string;
+    output?: string;
+    marker?: string;
     addMissingKeys?: boolean;
     replace?: boolean;
     defaultValue?: string | undefined;