+++ schematics/src/ng-add/index.ts
@@ -1,47 +1,35 @@
 import {
-  SchematicContext,
-  Tree,
-  Rule,
+  apply,
+  chain,
   EmptyTree,
   HostTree,
-  apply,
-  source,
-  move,
-  chain,
-  SchematicsException,
-  template,
-  url,
-  Source,
   mergeWith,
-  noop
+  move,
+  noop,
+  Rule,
+  SchematicContext,
+  SchematicsException,
+  source,
+  Source,
+  template,
+  Tree,
+  url
 } from '@angular-devkit/schematics';
-import { createSourceFile, SourceFile, ScriptTarget } from 'typescript';
+import { createSourceFile, ScriptTarget, SourceFile } from 'typescript';
 import { LIB_NAME } from '../schematics.consts';
-import { addImportToModule, insertImport, addProviderToModule } from '../utils/ast-utils';
+import { addImportToModule, addProviderToModule, insertImport } from '../utils/ast-utils';
 import { InsertChange } from '../utils/change';
 import { findRootModule } from '../utils/find-module';
 import { getProject, setEnvironments } from '../utils/projects';
-import { SchemaOptions, Loaders, TranslationFileTypes } from './schema';
+import { checkIfTranslationFilesExist } from '../utils/translations';
+import { createConfig } from '../utils/transloco';
+import { Loaders, SchemaOptions, TranslationFileTypes } from './schema';
 import { stringifyList } from '../utils/array';
 
 function jsonTranslationFileCreator(source, lang) {
   return source.create(
     `${lang}.json`,
-    `{
-  "title": "transloco ${lang}",
-  "dynamic": "transloco {{value}}"
-}
-`
-  );
-}
-
-function typescriptTranslationFileCreator(source, lang) {
-  return source.create(
-    `${lang}.ts`,
-    `export default {
-  title: "transloco ${lang}",
-  dynamic: "transloco {{value}}"
-};
+    `{}
 `
   );
 }
@@ -143,10 +131,7 @@ export default function(options: SchemaOptions): Rule {
     const isLib = project.projectType === 'library';
     const assetsPath = `${sourceRoot}/${options.path}`;
 
-    const translationCreator =
-      options.translateType === TranslationFileTypes.Typescript
-        ? typescriptTranslationFileCreator
-        : jsonTranslationFileCreator;
+    const translationCreator = jsonTranslationFileCreator;
 
     const translateFiles = apply(source(createTranslateFiles(langs, translationCreator)), [move('/', assetsPath)]);
 
@@ -166,8 +151,10 @@ export default function(options: SchemaOptions): Rule {
       updateEnvironmentBaseUrl(host, sourceRoot, 'http://localhost:4200');
     }
 
+    createConfig(host, langs, assetsPath);
+
     return chain([
-      mergeWith(translateFiles),
+      checkIfTranslationFilesExist(assetsPath, langs, '.json', true) ? noop() : mergeWith(translateFiles),
       options.loader === Loaders.Http
         ? chain([
             addImportsToModuleFile(options, ['HttpClientModule'], '@angular/common/http'),