+++ schematics/src/utils/transloco.ts
@@ -1,12 +1,35 @@
 import { PathFragment } from '@angular-devkit/core';
 import { DirEntry, Tree } from '@angular-devkit/schematics';
 import { getConfig as getTranslocoConfig, TranslocoConfig } from '@ngneat/transloco-utils';
+import { CONFIG_FILE } from '../schematics.consts';
+import { stringifyList } from './array';
 import { getProject } from './projects';
 
 const p = require('path');
 
 export function getConfig(): TranslocoConfig {
   return getTranslocoConfig();
+}
+
+export function createConfig(host: Tree, langs: string[], rootTranslationsPath = 'src/assets/i18n/') {
+  if (!host.get(CONFIG_FILE)) {
+    const config = `module.exports = {
+  rootTranslationsPath: '${rootTranslationsPath}',
+  langs: [${stringifyList(langs)}],
+  keysManager: {}
+};`;
+    host.create(CONFIG_FILE, config);
+  }
+}
+
+export function updateConfig(host: Tree, config: TranslocoConfig) {
+  const originalConfig = getConfig();
+  if (!originalConfig || Object.keys(originalConfig).length === 0) {
+    return createConfig(host, config.langs, config.rootTranslationsPath);
+  }
+  const stringifyConfig = JSON.stringify({ ...config, ...originalConfig }, null, 2);
+  const content = `module.exports = ${stringifyConfig};`;
+  host.overwrite(CONFIG_FILE, content);
 }
 
 export function getJsonFileContent(fileName: PathFragment, dir: DirEntry, parser = JSON.parse) {