+++ src/app/inline-loaders/inline/inline.component.ts
@@ -1,23 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { TRANSLOCO_SCOPE } from '@ngneat/transloco';
-
-export const loader = ['en', 'es'].reduce((acc, lang) => {
-  acc[lang] = () => import(`../i18n/${lang}.json`);
-  return acc;
-}, {});
 
 @Component({
   selector: 'app-inline',
-  templateUrl: './inline.component.html',
-  providers: [
-    {
-      provide: TRANSLOCO_SCOPE,
-      useValue: {
-        scope: 'inline',
-        loader
-      }
-    }
-  ]
+  templateUrl: './inline.component.html'
 })
 export class InlineComponent implements OnInit {
   constructor() {}